Completed implementation of OHM price lookup using Balancer
- Changed the definition of "base" tokens to be just stablecoins and wETH, now excluding OHM. This makes the price lookup more dynamic.
- Shifted "base" token price lookups into PriceBase.ts, to isolate it from other modules
- Fixed mock failures for tests, introduced as a result of the changes
