+++ .subgraph-version
@@ -1,1 +1,1 @@
-SUBGRAPH_VERSION=2.0.67
+SUBGRAPH_VERSION=2.0.68
